Extract out concern for completion endpoint.
