+++ services/QuillLMS/engines/evidence/lib/evidence/evidence/open_ai.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Evidence
+  module OpenAI
+    API_KEY = ENV['OPENAI_API_KEY']
+  end
+end
