+++ services/QuillLMS/engines/evidence/lib/evidence/evidence/open_ai/completion.rb
@@ -3,12 +3,11 @@
 module Evidence
   module OpenAI
     class Completion < Evidence::ApplicationService
-      include HTTParty
-      base_uri 'https://api.openai.com/v1'
+      include Evidence::OpenAI::APIConcern
 
-      API_KEY = ENV['OPENAI_API_KEY']
-      MAX_TOKENS = 24
       ENDPOINT = '/completions'
+
+      MAX_TOKENS = 24
       # A-D, (A)Least -> (D)Most Complex/Expensive
       MODELS = {
         ada: 'text-ada-001',
@@ -16,11 +15,11 @@ module Evidence
         curie: 'text-curie-001',
         davinci: 'text-davinci-002'
       }
-      BLANK = ''
+
       STOP_TOKENS = [". ", ", "]
       MAX_COUNT = 128 # API has an undocument max of 128 for 'n'
 
-      attr_accessor :response, :prompt, :temperature, :count, :model_key, :options_hash
+      attr_accessor :prompt, :temperature, :count, :model_key, :options_hash
 
       def initialize(prompt:, temperature: 0.5, count: 1, model_key: :babbage, options_hash: {})
         @prompt = prompt
@@ -30,11 +29,8 @@ module Evidence
         @options_hash = options_hash
       end
 
-      def headers
-        {
-          "Content-Type" => "application/json",
-          "Authorization" => "Bearer #{API_KEY}"
-        }
+      def endpoint
+        ENDPOINT
       end
 
       def request_body
@@ -46,39 +42,6 @@ module Evidence
           max_tokens: MAX_TOKENS,
           stop: STOP_TOKENS
         }.merge(options_hash)
-      end
-
-      def run
-        return response if response.present?
-
-        @response = self.class.post(ENDPOINT, body: request_body.to_json, headers: headers)
-
-        cleaned_results
-      end
-
-      def cleaned_results
-        result_texts_removed_characters
-          .map{|r| r&.split(/\n/)&.first } # drop anything after a \n
-          .map{|r| r&.strip } # remove leading/ending spaces
-          .compact
-          .select {|r| r.length >= 10}
-          .uniq
-      end
-
-      def result_texts_removed_characters
-        result_texts
-          .map{|r| r&.gsub(/^(\n|-|\s)+/, BLANK)} # strip all leading \n, -, or whitespace
-          .map{|r| r&.gsub(/(\]|\[|=|\d\))/, BLANK)} # strip brackets, equal signs, and 1), 2)
-      end
-
-      private def result_texts
-        response
-          .parsed_response['choices']
-          .map{|r| r['text']}
-      end
-
-      def reset_response
-        @response = nil
       end
     end
   end
